Organization of layers and sprites
I have organized and rearranged some of the layers and added a couple of more. I have divided the layers into three folders: "Floor plan", Props" and "Collisions". Additionally, I have added fe mor layers called "depth tiles 2" and "depth tiles 3" in addition to prop layers of foreground and background prop layers.

I have decorated the bone room and added few depth tiles to the lore hallway, passageway and exit rooms that previously weren't there.

The arrangement of the tile layers have unfortunately made the game to not work :/. But we have agreed upon a system to fix this. Using object layers to distinguish multiple tiles as one object. I know it'll be hell, but i think it would be worth it. Be better organize now or it will be a disaster later on in development.

Have a nice day,
Pedro.